% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{alloc_coef}
\alias{alloc_coef}
\title{Allocation coefficient matrix}
\usage{
alloc_coef(x, input_type = "industry", output_type = "industry")
}
\arguments{
\item{x}{An input-output table.}

\item{input_type}{Input type.}

\item{output_type}{Output type.}
}
\value{
A dibble of allocation coefficient matrix.
}
\description{
Allocation coefficient matrix
}
