% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotable.R
\name{as_iotable}
\alias{as_iotable}
\title{Input-output table}
\usage{
as_iotable(x, input = list(), output = list(), type = list())
}
\arguments{
\item{x}{A data frame.}

\item{input}{A named list of column names of inputs.}

\item{output}{A named list of column names of outputs.}

\item{type}{A named list of input or output names.}
}
\value{
A dibble of input-output table.
}
\description{
Input-output table
}
