% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{export_coef}
\alias{export_coef}
\title{Export coefficient vector}
\usage{
export_coef(x)
}
\arguments{
\item{x}{An input-output table.}
}
\value{
A dibble of export coefficient vector.
}
\description{
Export coefficient vector
}
