% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeff.R
\name{export_coeff}
\alias{export_coeff}
\title{Export coefficient vector}
\usage{
export_coeff(x)
}
\arguments{
\item{x}{An input-output table.}
}
\value{
A dibble of export coefficient vector.
}
\description{
Export coefficient vector
}
