% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{import_coef}
\alias{import_coef}
\title{Import coefficient vector}
\usage{
import_coef(x)
}
\arguments{
\item{x}{An input-output table.}
}
\value{
A dibble of import coefficient vector.
}
\description{
Import coefficient vector
}
