% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeff.R
\name{import_coeff}
\alias{import_coeff}
\title{Import coefficient vector}
\usage{
import_coeff(x)
}
\arguments{
\item{x}{An input-output table.}
}
\value{
A dibble of import coefficient vector.
}
\description{
Import coefficient vector
}
