% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{input_coef}
\alias{input_coef}
\title{Input coefficient matrix}
\usage{
input_coef(x, input_type = "industry", output_type = "industry")
}
\arguments{
\item{x}{An input-output table.}

\item{input_type}{Input type.}

\item{output_type}{Output type.}
}
\value{
A dibble of input coefficient matrix.
}
\description{
Input coefficient matrix
}
