% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iotable_sector13_2011}
\alias{iotable_sector13_2011}
\title{2011 input-output table for Japan (13 sectors)}
\format{
An object of class \code{ddf_col} with 19 rows and 23 columns.
}
\usage{
iotable_sector13_2011
}
\description{
2011 input-output table for Japan (13 sectors)
}
\keyword{datasets}
