% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inv.R
\name{leontief_inv}
\alias{leontief_inv}
\title{Leontief inverse matrix}
\usage{
leontief_inv(x, mat_type = c("I-(I-M)A", "I-A"))
}
\arguments{
\item{x}{An input-output table.}

\item{mat_type}{A type of Leontief inverse matrix.}
}
\value{
A dibble of Leontief inverse matrix.
}
\description{
Leontief inverse matrix
}
