% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{local_demand}
\alias{local_demand}
\title{Local demand}
\usage{
local_demand(x, input_type = "industry")
}
\arguments{
\item{x}{A input-output table.}

\item{input_type}{Input type.}
}
\value{
A dibble of local demand.
}
\description{
Local demand
}
