% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skyline.R
\name{production_inducement}
\alias{production_inducement}
\title{Production inducement}
\usage{
production_inducement(x, mat_type = c("open", "closed"))
}
\arguments{
\item{x}{An input-output table.}

\item{mat_type}{A type of Leontief inverse matrix.}
}
\value{
A dibble of production inducement.
}
\description{
Production inducement
}
