% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skyline.R
\name{self_sufficiency}
\alias{self_sufficiency}
\title{Self sufficiency}
\usage{
self_sufficiency(x, summary = FALSE)
}
\arguments{
\item{x}{An input-output table.}
}
\value{
A dibble of self sufficiency.
}
\description{
Self sufficiency
}
