% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect.R
\name{spillover_effect}
\alias{spillover_effect}
\title{Spillover effect}
\usage{
spillover_effect(x, finaldemand_change, mat_type = c("I-(I-M)A", "I-A"))
}
\arguments{
\item{x}{An input-output table.}

\item{finaldemand_change}{A named numeric vector of changes in final demand.}

\item{mat_type}{A type of Leontief inverse matrix.}
}
\value{
A dibble of spillover effect.
}
\description{
Spillover effect
}
