% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{total_input}
\alias{total_input}
\title{Total input}
\usage{
total_input(
  x,
  input_type = c("industry", "valueadded"),
  output_type = "industry"
)
}
\arguments{
\item{x}{A input-output table.}

\item{output_type}{Output type.}
}
\value{
A dibble of total input.
}
\description{
Total input
}
