% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value.R
\name{total_output}
\alias{total_output}
\title{Total output}
\usage{
total_output(
  x,
  input_type = "industry",
  output_type = c("industry", "finaldemand", "export", "import")
)
}
\arguments{
\item{x}{A input-output table.}

\item{input_type}{Input type.}

\item{output_type}{Output type.}
}
\value{
A dibble of total output.
}
\description{
Total output
}
